# cobol/specs/1-airline-management/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Airline Management System API
  version: 1.0.0
paths:
  /login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Unauthorized

  /flights:
    get:
      summary: Search flights
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
        - name: destination
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Flights retrieved successfully

    post:
      summary: Create a flight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
      responses:
        '201':
          description: Flight created successfully

  /bookings:
    post:
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Booking created successfully

components:
  schemas:
    Flight:
      type: object
      properties:
        origin:
          type: string
        destination:
          type: string
        departure_time:
          type: string
          format: date-time
        arrival_time:
          type: string
          format: date-time
        capacity:
          type: integer
        available_seats:
          type: integer

    Booking:
      type: object
      properties:
        flight_id:
          type: string
        customer_id:
          type: string
        seat_number:
          type: string